### 2019-04-26 #1413 @whedon
[PRE REVIEW]: Netwulf: Interactive visualization of networks in Python
**Submitting author:** @ulfaslak (<a href="http://orcid.org/0000-0003-4704-3609">Ulf Aslak</a>)
**Repository:** <a href="https://github.com/benmaier/netwulf" target ="_blank">https://github.com/benmaier/netwulf</a>
**Version:** 0.0.14
