### 2019-04-26 #1413 @danielskatz
👋 @ulfaslak - For JOSS, your submission must have a research application that is clearly defined in the README and paper.  Since I can't see your paper now, I'm just looking at the repo, but I don't see this there.  Please explain what the research application of this software is.

Also, please try to fix your paper, following [our example](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). When you have made changes and want to try to rebuild, enter a new comment here with ```@whedon generate pdf```
