### 2019-04-26 0b9e442
Update README.md
clarify research purpose
