### 2019-04-26 #1413 @ulfaslak
Hi @danielskatz.

> For JOSS, your submission must have a research application that is clearly defined in the README and paper. Since I can't see your paper now, I'm just looking at the repo, but I don't see this there. Please explain what the research application of this software is.

I updated the README with a clear statement about research purpose (in the About section).

> Also, please try to fix your paper, following our example.

Works now.
