### 2019-05-01 benmaier/netwulf#17 @leouieda
Add community guidelines
Hi @ulfaslak @benmaier, are there community guidelines for this project? For example, a contributing guide (usually `CONTRIBUTING.md`) and a code of conduct (usually `CODE_OF_CONDUCT.md` from [Contributor Covenant](https://www.contributor-covenant.org/)). I couldn't find anything in the repository or the documentation.

These are required for the JOSS submission (openjournals/joss-reviews#1425) and are usually straight forward to add. So I though I'd give you a head start before the review starts :slightly_smiling_face:
