### 2019-05-01 #1413 @leouieda
:wave: Hi @camillescott @vc1492a would you be interested in reviewing this submission for JOSS?
