### 2019-05-01 #1413 @camillescott
Sure thing, seems up my alley.

On Wed, May 1, 2019 at 12:46 PM Leonardo Uieda <notifications@github.com>
wrote:

> 👋 Hi @camillescott <https://github.com/camillescott> @vc1492a
> <https://github.com/vc1492a> would you be interested in reviewing this
> submission for JOSS?
>
