### 2019-05-02 #1425 @leouieda
@vc1492a @camillescott thank you for volunteering your time for this review :1st_place_medal: This is the review issue where you can leave comments and check items from your checklists. If you open any issues or PRs in the software repository, please remember to link back to this issue using `openjournals/joss-reviews#1425` in the issue/PR so we can keep track of progress.

@ulfaslak @vc1492a @camillescott please don't hesitate to reach to me if you have questions/concerns (here or privately, if you prefer).
