### 2019-05-06 #1425 @vc1492a
**Overview**: This was an interesting piece of software to review and I enjoyed walking through it and testing its functionality. While I believe updates are needed in order for publication in JOSS (see below), the software is well documented and seems to fit a niche that may be useful within the network analysis community.

**License**: While a license is provided in the project repository (MIT), it may be beneficial to indicate the license more publicly in the readme on Github and/or in the documentation. For example, a badge could be used or a section added to the readme or documentation.

**Version Matching**: The version number submitted to JOSS does match the latest version of the source code in master on Github. It may be beneficial however to tag the releases on Github that are uploaded to PyPi such that users can download the software directly from Github via a Release as opposed to having to clone or pull the repository. This is helpful to users in some cases and also provides the developer with a copy of the source code for previous versions.

**Functionality**: I was able to verify the basic functionality, in this case installing the software, importing the software, and opening the visualization interface successfully by calling `netwulf.visualize(G)`, where `G` is a networkX graph object. I was also able to verify the download PNG functionality as well as the CSV upload functionality using the example provided in the documentation. However, I was not able to get the `Post to Python` functionality working. More specifically, the UI indicates `success` and closes the browser window, but the Python kernel remains busy indefinitely and on manually stopping execution, the `network_properties` object is not returned. This occurred in both a Jupyter notebook as well as a terminal window. Consider opening an issue with the tag `bug`.

I'm not sure, but I think there may be a CSS formatting or similar error in the user interface. See the screenshot below:

![Upload Text](https://i.imgur.com/ueLnh0C.png)

The `Upload file` section open parenthesis and does not close them - seems to be missing ` json)`. No big deal, but may be worth addressing in a future release. Consider opening an issue with the tag `bug`.

Some potentially useful functionality would be to allow a user to start the visualization interface without the need to pass a graph object into `visualize()`. This may be useful in the case where the user has a file that they wish to examine and want to upload, but may not want to load first within the Python environment. By making the graph object an optional argument, the user could start the visualization interface and then explore one or more files through the upload functionality. This would be a nice feature in future releases but isn't necessary to address as part of this review. Consider opening an issue with the tag `enhancement`.

Additionally, I noticed a lot of use of Python's `os` library. While this works fine on Linux and Unix devices, it can result in errors on Windows machines. Using `Pathlib` in place of `os` would reduce the likelihood of errors on Windows machines. I am not able to verify the functionality of this software on Windows machines (the editor can decide whether this functionality on windows should be formally verified or whether this potential limitation should be listed in `readme.md` or documentation), but do know that `Pathlib` would allow the developers to normalize paths for Windows machines easily. Consider opening an issue with the tag `enhancement`.

**Performance claims**: The paper states "The interactive visualization is implemented in JavaScript, relies on d3.js for computing layouts, and uses the HTML5-object canvas for rendering. This makes it, to our knowledge, the most performant tool for interactive network visualization in Python to date." This statement implies that the word `performant` is used to describe the speed in which one can create and manipulate a network graph and return an object to Python that describes that graph. After using the software I believe this claim is true, but would suggest the editor or others also look into this area to help verify this claim.

**Functionality documentation**: This documentation is great and contains details on the API as well as cookbooks containing example code. I was able to use the example code easily to get started on verifying the functionality of the software. I noticed one small item that could be addressed in the examples. In the below screenshot, it appears that a necessary `matplotlib` import is missing:

![Missing Matplotlib](https://i.imgur.com/3TCJC4K.png)

I believe this should include `import matplotlib.pyplot as plt`.

Additionally, in the example located [in the filtering section](https://netwulf.readthedocs.io/en/latest/python_api/network_manipulation.html ), `grp = {u: 'ABCDE'[u%5]  for u in G.nodes()}` and `nx.set_node_attributes(G, grp, 'wum')` should be executed after the first visualization and before the second visualization, as opposed to before the first. Executing those lines prior to the first example results in the following error when opening the visualization interface:

![Example Error](https://i.imgur.com/Oet9mDX.png)

The rest of the documentation looks great and is quite helpful! To gain additional feedback and user adoption, it may be interesting to show some insights gained through using the interactive visualization interface on some real-world data (not necessary though).

**Statement of need (documentation)**: A statement of need exists in the documentation, but the focus differs from that highlighted in the generated paper. The documentation states "Netwulf is an interactive visualization tool for networkx Graph-objects, that allows you to produce beautiful looking network visualizations. Simply input a networkx.Graph object, style the network in the interactive console and either download the result as a PNG or pipe the layout back to Python for further processing." This implies the software's focus is on creating visually-pleasing network visualizations easily from networkx objects. However, the paper states "Netwulf is a light-weight Python library that provides a simple API for interactively visualizing a network and returning the computed layout and style. It is build around the philosophy that network manipulation and preprocessing should be done programmatically, but that the efficient generation of a visually appealing network is best done interactively, without code. Therefore, it offers no analysis functionality and only few exploration features, but instead focuses almost entirely on fast and intuitive layout manipulation and node/link styling." This statement implies that the development focus is still in generating visually appealing networks interactively, but is mentions a programmatic interface and lack of analysis tools which obfuscates that meaning somewhat. The two statements are not in direct conflict, but the purpose of the work must be clearly stated and in agreement between the paper and the documentation and should be revised for clarity.

**Installation instructions**: Instructions for installation work and are easy to follow, with the listed dependencies in `setup.py` being installed upon installation. However, these dependencies are in the documentation and do not appear on the Github readme and should be mentioned in both locations for clarity.

**Automated tests**: I did not see any automated tests in this repository that could be used to verify the functionality of the software. Revisions of this work should include a series of unit tests that can be used to verify the functionality of the software, through testing of function outputs, input and output formatting, etc. There are a variety of ways to write and use unit tests in Python, two of which are the `pytest` and `pytest-cov` packages - the former allows developers to execute unit tests (with helpful extensions for writing tests, too) while the latter provides functionality to `pytest` for coverage statistics, e.g. which lines of code your unit tests cover and the percentage of lines in each file (and overall, the package) that is covered by the written unit tests (this coverage statistic could be reported in `readme.md`). Additionally, if you're looking to have others contribute to the project more easily in the future consider setting up a continuous integration pipeline that will automatically run tests and report the test results (e.g. using Travis CI) and coverage (e.g. using Coveralls) with any pushes to `master` or `dev` and for any pull requests opened by open-source contributors. This pipeline would ease the burden on making sure new additions and/or changes to software work as intended, and would also allow maintainers to formally test against every version of Python 3 that should be supported. This would allow the authors to state "this software works in Python versions 3.5 - 3.7" as opposed to simply stating "works in Python 3".

**Community guidelines**: While there are some development notes on how to update the front-end from it's upstream directory when developing, there are no clear guidelines on Github (in the readme or otherwise) or in the documentation on how to best jump into development or open a pull request for bug fixes or new features. Additionally, there is no documentation on how to report issues and it's implicitly implied that the user should open a Github issue. If that's the desired error reporting behavior, this should be made more explicit in the documentation. If Github issues is the desired mechanism, it may be beneficial to the developers to open issues for some of the behaviors noted here - for the authors: let me know if you'd like me to open issues for the items identified in the `Functionality` section above.

**References**: `paper.bib` had all relevant references, but did not have all DOIs that are available, so added a few and opened a [this pull request](https://github.com/benmaier/netwulf/pull/18). Specifically for the `webweb` reference, the DOI is pending since it's currently under review by JOSS (https://github.com/openjournals/joss-reviews/issues/1308). The editor of this work should decide how to handle this reference and whether updating the `paper.bib` file post publication in JOSS is a possibility so that this reference's DOI may be included for this work once itself is reviewed and later published.

**Other**: Although the visuals current used are already nice and informative, they could also be improved in future versions. This would (potentially) be a nice place to collaborate with others with, and documentation on how to contribute to this and other areas would be nice to see in the repository and/or documentation. Also consider adding a `changelog.md` file that details changes, additions, and fixes made to the code with each release version. Lastly, it seems that [semantic versioning](https://semver.org/) is used but this isn't make explicitly clear in `readme.md` - not that it needs to be, but adding details such as these, along with a `changelog.md`, would go a long way towards easing the entry into development for those which may want to contribute to the package.
