### 2019-05-07 #1425 @leouieda
@vc1492a thank you for the detailed review :rocket: @ulfaslak please do your best to respond to Valentino's comments either here in this thread or in separate issues in your repository (preferred). If you open the issues/PRs, please remember to link back to this thread by including `openjournals/joss-reviews#1425` in the issue/PR body.
