### 2019-05-07 benmaier/netwulf#19 @leouieda
Add the required Python versions to setup.py
It's best to specify the compatible Python versions in the `setup.py` using the [`python_requires`](https://packaging.python.org/guides/distributing-packages-using-setuptools/#python-requires) argument. This let's pip et al know if your Python version is compatible and avoids installing broken packages.

It will probably be useful if adopting `Pathlib` which has some features that are 3.6+ only. I don't see many reasons to use 3.5 since most of the scientific stack has moved on to 3.6.

openjournals/joss-reviews#1425
