### 2019-05-10 #1425 @benmaier
@vc1492a, indeed thank you very much for the super detailed review and @leouieda thank you for the thorough editing process so far. Since @ulfaslak and me are both on vacation at the moment, we will be responding to the review in detail afterwards. In the mean time, @vc1492a would you be so kind as to open an issue for the `Functionality` problem that you encountered? Indeed, we had a similar problem before (see https://github.com/benmaier/netwulf/issues/4 and https://github.com/benmaier/netwulf/issues/15) but thought we fixed the bug you unfortunately encountered. Thank you very much!
