### 2019-05-13 8a5a5e4
fixed formula in comments
