### 2019-05-15 benmaier/netwulf#21 @benmaier
Let `visualize` be run without argument such that the interface opens and a file can be loaded manually
as suggested by @vc1492a in https://github.com/openjournals/joss-reviews/issues/1425

> Some potentially useful functionality would be to allow a user to start the visualization interface without the need to pass a graph object into visualize(). This may be useful in the case where the user has a file that they wish to examine and want to upload, but may not want to load first within the Python environment. By making the graph object an optional argument, the user could start the visualization interface and then explore one or more files through the upload functionality. This would be a nice feature in future releases but isn't necessary to address as part of this review. Consider opening an issue with the tag enhancement.
