### 2019-05-15 benmaier/netwulf#22 @benmaier
bugs in documentation code
as remarked by @vc1492a in openjournals/joss-reviews#1425

> I noticed one small item that could be addressed in the examples. In the below screenshot, it appears that a necessary matplotlib import is missing:

![68747470733a2f2f692e696d6775722e636f6d2f3354434a43344b2e706e67](https://user-images.githubusercontent.com/10728380/57755521-709c5400-76f1-11e9-86ba-a5fbc964e18a.png)

> I believe this should include import matplotlib.pyplot as plt.

> Additionally, in the example located in the filtering section, grp = {u: 'ABCDE'[u%5] for u in G.nodes()} and nx.set_node_attributes(G, grp, 'wum') should be executed after the first visualization and before the second visualization, as opposed to before the first. Executing those lines prior to the first example results in the following error when opening the visualization interface:

![68747470733a2f2f692e696d6775722e636f6d2f4f6574396d44582e706e67](https://user-images.githubusercontent.com/10728380/57755542-7a25bc00-76f1-11e9-923e-89f1c8f38bdc.png)
