### 2019-05-15 benmaier/netwulf#23 @benmaier
Switch to `Pathlib` and abandon `os.path`
as discussed in https://github.com/openjournals/joss-reviews/issues/1425

> Additionally, I noticed a lot of use of Python's os library. While this works fine on Linux and Unix devices, it can result in errors on Windows machines. Using Pathlib in place of os would reduce the likelihood of errors on Windows machines. I am not able to verify the functionality of this software on Windows machines (the editor can decide whether this functionality on windows should be formally verified or whether this potential limitation should be listed in readme.md or documentation), but do know that Pathlib would allow the developers to normalize paths for Windows machines easily. Consider opening an issue with the tag enhancement.

See also #19
