### 2019-05-15 benmaier/netwulf#19 @benmaier
see also this comment in openjournals/joss-reviews#1425

> Installation instructions: Instructions for installation work and are easy to follow, with the listed dependencies in setup.py being installed upon installation. However, these dependencies are in the documentation and do not appear on the Github readme and should be mentioned in both locations for clarity.
