### 2019-05-15 benmaier/netwulf#25 @benmaier
Add changelog
see openjournals/joss-reviews#1425

> consider adding a changelog.md file that details changes, additions, and fixes made to the code with each release version. Lastly, it seems that semantic versioning is used but this isn't make explicitly clear in readme.md - not that it needs to be, but adding details such as these, along with a changelog.md, would go a long way towards easing the entry into development for those which may want to contribute to the package.
