### 2019-05-15 80b2776
added code of conduct, see #17
