### 2019-05-15 a861825
added contributing statement
