+++ CONTRIBUTING.md
@@ -1,0 +1,21 @@
+# Contributing
+
+When contributing to this repository, please first discuss the change you wish to make via an [issue](https://github.com/benmaier/netwulf/issues/new).
+
+Please note we have a [code of conduct](https://github.com/benmaier/netwulf/raw/master/CONDE_OF_CONDUCT.md), please follow it in all your interactions with the project.
+
+## Pull Request Process
+
+1. Ensure any install or build dependencies are removed before the end of the layer when doing a 
+   build.
+2. Update README.md and CHANGELOG.md with details of changes to the interface, this includes new environment 
+   variables, exposed ports, useful file locations and container parameters.
+3. Increase the version number in [netwulf/metadata.py](https://github.com/benmaier/netwulf/blob/master/netwulf/metadata.py) and the README.md to the new version that this
+   Pull Request would represent. The versioning scheme we use is [SemVer](http://semver.org/).
+4. You may merge the Pull Request in once you have the sign-off of one other developer, or if you 
+   do not have permission to do that, you may request the reviewer to merge it for you.
+
+## Attribution
+
+This Contributing Statement is adapted from [this template](https://gist.github.com/PurpleBooth/b24679402957c63ec426) by @PurpleBooth.
+
