+++ README.md
@@ -1,3 +1,5 @@
+[![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-v1.4%20adopted-ff69b4.svg)](code-of-conduct.md)
+
 ![logo](https://github.com/benmaier/netwulf/raw/master/img/logo_small.png)
 
 ## About
@@ -7,20 +9,30 @@
 ![example](https://github.com/benmaier/netwulf/raw/master/img/simple_example.gif)
 
 
-
 ## Install
 
     pip install netwulf
 
-Beware: `netwulf` only works with Python 3!
+`netwulf` was developed and tested for 
+
+* Python 3.5
+* Python 3.6
+* Python 3.7
+
+## Dependencies
+
+`netwulf` directly depends on the following packages which will be installed by `pip` during the installation process
+
+* `networkx>=2.0`
+* `numpy>=0.14`
+* `matplotlib>=3.0`
+* `simplejson>=3.0`
 
 ## Documentation
 
-The full documentation is available [here](https://netwulf.readthedocs.io/en/latest/python_api/start.html).
+The full documentation is available at https://netwulf.rtfd.io .
 
 ## Example
-
-### Standard
 
 Create a network and look at it
 
@@ -34,120 +46,11 @@ visualize(G)
 
 ![visualization example0](https://github.com/benmaier/netwulf/raw/master/img/BA_1.png)
 
-When the visualization was posted to Python in the Browser, the function actually returns a dictionary containing all style information of the stylized network. For instance,
+## Contributing
 
-```python
-import networkx as nx
-from netwulf import visualize
+If you want to contribute to this project, please make sure to read the [code of conduct](https://github.com/benmaier/netwulf/raw/master/CONDE_OF_CONDUCT.md) and the [contributing guidelines](https://github.com/benmaier/netwulf/raw/master/CONTRIBUTING.md).
 
-G = nx.barabasi_albert_graph(5,m=1)
-# the following only works if the user clicked "Post to Python" in the visualization.
-network_properties, config = visualize(G)
-print(network_properties)
-```
-
-A possible result from this code is
-
-```python
-{   'height': 919,
-    'linkAlpha': 0.5,
-    'linkColor': '#7c7c7c',
-    'links': [   {'link': [0, 1], 'width': 2},
-                 {'link': [0, 2], 'width': 2},
-                 {'link': [0, 3], 'width': 2},
-                 {'link': [2, 4], 'width': 2}],
-    'nodeStrokeColor': '#000000',
-    'nodeStrokeWidth': 0.5,
-    'nodes': [   {   'color': '#16a085',
-                     'id': 0,
-                     'pos': [442.2593240813363, 454.37557840980224],
-                     'radius': 5.632111911473014},
-                 {   'color': '#16a085',
-                     'id': 1,
-                     'pos': [481.8171665449027, 438.58151881303337],
-                     'radius': 5.632111911473014},
-                 {   'color': '#16a085',
-                     'id': 2,
-                     'pos': [446.98116529402097, 495.93002968604276],
-                     'radius': 5.632111911473014},
-                 {   'color': '#16a085',
-                     'id': 3,
-                     'pos': [437.9340114428489, 414.42238247014905],
-                     'radius': 5.632111911473014},
-                 {   'color': '#16a085',
-                     'id': 4,
-                     'pos': [488.5033653510982, 494.1926029578719],
-                     'radius': 5.632111911473014}],
-    'width': 919
-}
-```
-
-
-### Config
-
-It's possible to change the default settings which are
-
[…]
