### 2019-05-15 9deefab
added python_requires, see #19
