### 2019-05-15 benmaier/netwulf#17 @benmaier
Added. We should further discuss whether to add something about contributing in the documentation, too.
