### 2019-05-15 66a5210
fix example bugs, close #22
