### 2019-05-15 benmaier/netwulf#21 @benmaier
I think I disagree here. Loading a network from a file in Python takes usually [a single line of code using networkx](https://networkx.github.io/documentation/latest/auto_examples/basic/plot_read_write.html#sphx-glr-auto-examples-basic-plot-read-write-py) which should be simple enough and lines up with the philosophy that everything besides styling the network should be done programmatically. In fact, it seems that version 0.0.14 on PyPI contained outdated code which actually let the user upload files. In the actual (now updated version 0.0.16), these parts of the UI are suppressed in the interface started in Python and only shown in its online version (see https://ulfaslak.com/works/network_styling_with_d3/index.html). I apologize for the mismatch of the interface in the PyPI-version and the latest github-commit. This shouldn't have happened.
