+++ CHANGELOG.md
@@ -1,0 +1,17 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+## [0.0.16] - 2019-05-15
+### Added
+- [code of conduct](https://github.com/benmaier/netwulf/raw/master/CONDE_OF_CONDUCT.md)
+- [contributing guidelines](https://github.com/benmaier/netwulf/raw/master/CONTRIBUTING.md).
+
+[Unreleased]: https://github.com/benmaier/netwulf/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/benmaier/netwulf/compare/v0.0.16...v1.0.0
+[0.0.16]: https://github.com/benmaier/netwulf/releases/tag/v0.0.16
