+++ README.md
@@ -46,6 +46,10 @@ visualize(G)
 
 ![visualization example0](https://github.com/benmaier/netwulf/raw/master/img/BA_1.png)
 
+## Changelog
+
+Changes are logged in a [separate file](https://github.com/benmaier/netwulf/raw/master/CHANGELOG.md).
+
 ## Contributing
 
 If you want to contribute to this project, please make sure to read the [code of conduct](https://github.com/benmaier/netwulf/raw/master/CONDE_OF_CONDUCT.md) and the [contributing guidelines](https://github.com/benmaier/netwulf/raw/master/CONTRIBUTING.md).
