### 2019-05-15 benmaier/netwulf#21 @benmaier
tagging @ulfaslak to ask for comment
