### 2019-05-15 f138acd
added contribution notes to the docs, close #17
