### 2019-05-15 4e054d7
clarify statement, close #24
