### 2019-05-15 benmaier/netwulf#24 @benmaier
I agree that both statements are not in conflict but have a somewhat different focus. I changed the statement in the documentation to

```
Netwulf is an interactive visualization tool for networkx_ Graph-objects,
that allows you to produce beautifully looking network visualizations. Simply
input a `networkx.Graph` object, style the network in the interactive console
and either download the result as a PNG or pipe the layout back to Python for
further processing. Netwulf is fast and relies on no crude dependencies.
It is build around the philosophy that network manipulation and preprocessing
should be done programmatically, but that the efficient generation of a visually
appealing network is best done interactively, without code.
```

and added a similar note in the README.md.

It is important to us to make it clear that we do not want `netwulf` to be another `Gephi` (used for analysis and exploration) but that its main focus is interactive visualization.
