### 2019-05-15 benmaier/netwulf#26 @benmaier
My main problem with this point is that the main purpose of the package is the interactive user-interface. I'm a bit lost on how to introduce automated testing for user interaction. @ulfaslak comments?
