+++ netwulf/metadata.py
@@ -3,10 +3,10 @@
 Contains a bunch of information about this package.
 """
 
-__version__ = "0.0.16"
+__version__ = "0.0.17"
 
 __author__ = "Ulf Aslak, Benjamin F. Maier"
-__copyright__ = "Copyright 2018, Ulf Aslak, Benjamin F. Maier"
+__copyright__ = "Copyright 2018-2019, Ulf Aslak, Benjamin F. Maier"
 __credits__ = ["Benjamin F. Maier"]
 __license__ = "MIT"
 __maintainer__ = "Benjamin F. Maier"
