+++ README.rst
@@ -1,13 +1,22 @@
 |logo|
 
-netwulf
-=======
+About
+-----
 
-This package provides an interface between
-`networkx <https://networkx.github.io/>`__ Graph objects and `Ulf
-Aslak's interactive web
-app <https://github.com/ulfaslak/network_styling_with_d3>`__ for simple
-and better network visualizations.
+**Simple and interactive network visualization in Python.** Network
+visualization is an indispensable tool for exploring and communicating
+patterns in complex systems. Netwulf offers an ultra-simple API for
+**reproducible interactive visualization** of networks directly from a
+Python prompt or Jupyter notebook. As a research tool, its purpose is to
+allow hassle-free quick interactive layouting/styling for communication
+purposes.
+
+The package is build around the philosophy that network manipulation and
+preprocessing should be done programmatically, but that the efficient
+generation of a visually appealing network is best done interactively,
+without code.
+
+|example|
 
 Install
 -------
@@ -16,13 +25,30 @@ Install
 
     pip install netwulf
 
-Beware: ``netwulf`` only works with Python 3!
+``netwulf`` was developed and tested for
+
+-  Python 3.5
+-  Python 3.6
+-  Python 3.7
+
+Dependencies
+------------
+
+``netwulf`` directly depends on the following packages which will be
+installed by ``pip`` during the installation process
+
+-  ``networkx>=2.0``
+-  ``numpy>=0.14``
+-  ``matplotlib>=3.0``
+-  ``simplejson>=3.0``
+
+Documentation
+-------------
+
+The full documentation is available at https://netwulf.rtfd.io .
 
 Example
 -------
-
-Standard
-~~~~~~~~
 
 Create a network and look at it
 
@@ -36,123 +62,22 @@ Create a network and look at it
 
 |visualization example0|
 
-When the visualization was posted to Python in the Browser, the function
-actually returns a dictionary containing all style information of the
-stylized network. For instance,
+Changelog
+---------
 
-.. code:: python
+Changes are logged in a `separate
+file <https://github.com/benmaier/netwulf/blob/master/CHANGELOG.md>`__.
 
-    import networkx as nx
-    from netwulf import visualize
+Contributing
+------------
 
-    G = nx.barabasi_albert_graph(5,m=1)
-    # the following only works if the user clicked "Post to Python" in the visualization.
-    network_properties, config = visualize(G)
-    print(network_properties)
+If you want to contribute to this project, please make sure to read the
+`code of
+conduct <https://github.com/benmaier/netwulf/blob/master/CODE_OF_CONDUCT.md>`__
+and the `contributing
+guidelines <https://github.com/benmaier/netwulf/blob/master/CONTRIBUTING.md>`__.
 
-A possible result from this code is
-
-.. code:: python
-
-    {   'height': 919,
-        'linkAlpha': 0.5,
-        'linkColor': '#7c7c7c',
-        'links': [   {'link': [0, 1], 'width': 2},
-                     {'link': [0, 2], 'width': 2},
-                     {'link': [0, 3], 'width': 2},
-                     {'link': [2, 4], 'width': 2}],
-        'nodeStrokeColor': '#000000',
-        'nodeStrokeWidth': 0.5,
-        'nodes': [   {   'color': '#16a085',
-                         'id': 0,
-                         'pos': [442.2593240813363, 454.37557840980224],
-                         'radius': 5.632111911473014},
-                     {   'color': '#16a085',
-                         'id': 1,
-                         'pos': [481.8171665449027, 438.58151881303337],
-                         'radius': 5.632111911473014},
-                     {   'color': '#16a085',
[…]
