+++ setup.cfg
@@ -2,4 +2,6 @@
 long-description = file: README.rst
 license = MIT
 classifiers =
-    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.5
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
