### 2019-05-15 baddcb4
switched to pathlib, fix #23
