+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.17] - 2019-05-15
+### Changed
+- switched from usage of `os.path` to `pathlib` at the appropriate places
+
 ## [0.0.16] - 2019-05-15
 ### Added
 - [code of conduct](https://github.com/benmaier/netwulf/raw/master/CONDE_OF_CONDUCT.md)
 - [contributing guidelines](https://github.com/benmaier/netwulf/raw/master/CONTRIBUTING.md).
 
-[Unreleased]: https://github.com/benmaier/netwulf/compare/v1.0.0...HEAD
-[1.0.0]: https://github.com/benmaier/netwulf/compare/v0.0.16...v1.0.0
+[Unreleased]: https://github.com/benmaier/netwulf/compare/v0.0.17...HEAD
+[0.0.17]: https://github.com/benmaier/netwulf/compare/v0.0.16...v0.0.17
 [0.0.16]: https://github.com/benmaier/netwulf/releases/tag/v0.0.16
