### 2019-05-15 bbc2ccf
License note as demanded in review process
