### 2019-05-15 ccea6d1
added OS notes
