### 2019-05-15 benmaier/netwulf#21 @ulfaslak
I'm with @benmaier on this one. It's a nice idea, and I'm happy you bring it up, because it does offer insight into what interactions feel intuitive, but I think it's going to create more confusion than utility. Moreover, the `visualize` function can actually take a dictionary object as input, so if one had such an object stored locally, the better way to visualize it in `netwulf` would be to load and serve it as input to `visualize`.
