### 2019-05-15 benmaier/netwulf#26 @ulfaslak
I see the point, especially for making contribution easier. The problem here is that the key functionality really has no immediate output, at least not without user interaction. To write automated tests, we would have to fill in the interaction step with something like `selenium` webdriver, but I can easily see that spinning out of proportion and introducing more obscurity than validation.
