### 2019-05-15 benmaier/netwulf#21 @vc1492a
@benmaier @ulfaslak Thanks for the commentary here - as you both mentioned, it looks like the newer version no longer allows users to upload files themselves in the interface and thus makes sense to require the user to pass some sort of object to initialize the visualization interface.
