+++ setup.py
@@ -24,6 +24,13 @@ setup(
                 'matplotlib>=3.0',
                 'simplejson>=3.0',
             ],
+    tests_require=['pytest', 'pytest-cov'],
+    setup_requires=['pytest-runner'],
+    classifiers=['License :: OSI Approved :: MIT License',
+           'Programming Language :: Python :: 3.5',
+           'Programming Language :: Python :: 3.6',
+           'Programming Language :: Python :: 3.7'
+           ],
     include_package_data = True,
     zip_safe = False,
 )
