+++ setup.cfg
@@ -1,3 +1,7 @@
+[aliases]
+# required for python setup.py test 
+test = pytest
+
 [metadata]
 long-description = file: README.rst
 license = MIT
