+++ .travis.yml
@@ -19,3 +19,11 @@ python:
   - "3.5"
   - "3.6"
   - "3.7"
+
+# Handle git submodules yourself
+git:
+    submodules: false
+# Use sed to replace the SSH URL with the public URL, then initialize submodules
+before_install:
+    - sed -i 's/git@github.com:/https:\/\/github.com\//' .gitmodules
+    - git submodule update --init --recursive
