+++ setup.py
@@ -25,16 +25,16 @@ setup(
     tests_require=['pytest', 'pytest-cov'],
     setup_requires=['pytest-runner'],
     classifiers=['License :: OSI Approved :: MIT License',
-           'Programming Language :: Python :: 3.5',
-           'Programming Language :: Python :: 3.6',
-           'Programming Language :: Python :: 3.7'
-           ],
+                'Programming Language :: Python :: 3.5',
+                'Programming Language :: Python :: 3.6',
+                'Programming Language :: Python :: 3.7'
+               ],
     project_urls={
-    'Documentation': 'https://netwulf.rtfd.io',
-    'Contributing Statement': 'https://github.com/benmaier/netwulf/blob/master/CONTRIBUTING.md',
-    'Bug Reports': 'https://github.com/benmaier/netwulf/issues',
-    'Source': 'https://github.com/benmaier/netwulf/',
-    'PyPI': 'https://pypi.org/project/netwulf/',
+        'Documentation': 'https://netwulf.rtfd.io',
+        'Contributing Statement': 'https://github.com/benmaier/netwulf/blob/master/CONTRIBUTING.md',
+        'Bug Reports': 'https://github.com/benmaier/netwulf/issues',
+        'Source': 'https://github.com/benmaier/netwulf/',
+        'PyPI': 'https://pypi.org/project/netwulf/',
     },
     include_package_data = True,
     zip_safe = False,
