### 2019-05-16 77e993f
catch fileDoesNotExist for Python 3.5
