+++ tests/test_tools.py
@@ -1,0 +1,150 @@
+import unittest
+
+import numpy as np
+import matplotlib.pyplot as pl
+import networkx as nx
+
+from netwulf.tools import bind_positions_to_network, get_filtered_network, draw_netwulf
+from netwulf import visualize
+
+def _get_test_network():
+    G = nx.Graph()
+    G.add_nodes_from(range(2))
+    G.add_nodes_from("ab")
+    G.add_edges_from([("a","b")])
+
+    return G
+
+def _round_positions(props):
+
+    for i, n in enumerate(props['nodes']):
+        props['nodes'][i]['x'] = np.around(props['nodes'][i]['x'],2)
+        props['nodes'][i]['y'] = np.around(props['nodes'][i]['y'],2)
+
+class Test(unittest.TestCase):
+
+    maxDiff = None
+
+    def test_posting(self):
+        """Test whether results are sucessfully posted to Python."""
+        G = _get_test_network()
+        visualize(G,is_test=True)
+
+    def test_reproducibility(self):
+        """Test whether a restarted network visualization with binded positions results in the same visualization."""
+        G = _get_test_network()
+        props, config = visualize(G,is_test=True)        
+
+        bind_positions_to_network(G, props)
+        newprops, newconfig = visualize(G, config=config,is_test=True)
+
+        # round all positional values to their 4th decimal place
+        _round_positions(props)
+        _round_positions(newprops)
+
+        # the second visualization should have frozen the nodes
+        assert(newconfig['freeze_nodes'] == True)
+
+        # change it so the dictionary-equal-assertion doesnt fail
+        newconfig['freeze_nodes'] = False
+
+        self.assertDictEqual(props, newprops)
+        self.assertDictEqual(config, newconfig)
+    
+    def test_filtering(self):
+        """Test whether filtering works the way it should."""
+
+        G = _get_test_network()
+        for u, v in G.edges():
+            G[u][v]['foo'] = np.random.rand()
+            G[u][v]['bar'] = np.random.rand()
+
+        grp = {u: 'AB'[i%2]  for i, u in enumerate(G.nodes()) }
+
+        new_G = get_filtered_network(G,edge_weight_key='foo')
+        visualize(new_G,is_test=True)
+
+        nx.set_node_attributes(G, grp, 'wum')
+
+        new_G = get_filtered_network(G,edge_weight_key='bar',node_group_key='wum')
+        visualize(new_G,is_test=True)
+
+    def test_matplotlib(self):
+        stylized_network = {   'linkAlpha': 0.5,
+                               'linkColor': '#7c7c7c',
+                               'links': [   {'source': 0, 'target': 1, 'width': 8.067180430164552},
+                                            {'source': 0, 'target': 2, 'width': 8.067180430164552},
+                                            {'source': 0, 'target': 4, 'width': 8.067180430164552},
+                                            {'source': 0, 'target': 6, 'width': 8.067180430164552},
+                                            {'source': 1, 'target': 3, 'width': 8.067180430164552},
+                                            {'source': 1, 'target': 7, 'width': 8.067180430164552},
+                                            {'source': 4, 'target': 5, 'width': 8.067180430164552},
+                                            {'source': 4, 'target': 9, 'width': 8.067180430164552},
+                                            {'source': 6, 'target': 8, 'width': 8.067180430164552}],
+                               'nodeStrokeColor': '#000000',
+                               'nodeStrokeWidth': 2.898890104734921,
+                               'nodes': [   {   'color': '#16a085',
+                                                'id': 0,
+                                                'radius': 27.71774321528189,
+                                                'x': 425.30610641802105,
+                                                'y': 408.2306291226664},
+                                            {   'color': '#16a085',
+                                                'id': 1,
[…]
