### 2019-05-16 benmaier/netwulf#26 @benmaier
added testing functionality
