### 2019-05-17 17765d6
include tests in package and update readme with dev notes
