+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
-testpaths = tests/
-python_files = tests/*.py
+testpaths = netwulf/tests/
+python_files = netwulf/tests/*.py
 python_class = Test
 pep8ignore = E221,E226,E241,E242
 addopts= --cov netwulf/
