### 2019-05-17 b4bb624
added dev notes section
