+++ sandbox/test_testing.py
@@ -1,0 +1,13 @@
+from netwulf.tests import Test
+
+T = Test()
+
+T.test_posting()
+
+T.test_filtering()
+
+T.test_matplotlib()
+
+T.test_reproducibility()
+
+
