### 2019-05-17 a305733
fix bug causing nx.Graphs with node positions to render in whole canvas domain
