### 2019-05-17 c1d5860
fix bug causing nx.Graphs with node positions to render in whole canvas domain
