### 2019-05-17 b4bef4a
matplotlib renders special *_canvas coordinates; set singleton nodes default true
