### 2019-05-17 a68194d
fix silly code error in docs
