### 2019-05-23 b7205f4
bugfixes and better testing
