+++ netwulf/tests/test_all.py
@@ -4,7 +4,7 @@ import numpy as np
 import matplotlib.pyplot as pl
 import networkx as nx
 
-from netwulf.tools import bind_positions_to_network, get_filtered_network, draw_netwulf
+from netwulf.tools import bind_properties_to_network, get_filtered_network, draw_netwulf
 from netwulf import visualize
 
 def _get_test_network():
@@ -19,11 +19,22 @@ def _get_test_network():
 def _get_test_config():
     return {'node_size':2,'link_width':2}
 
-def _round_positions(props):
+def _drastically_round_positions(props,nprec=-1):
 
     for i, n in enumerate(props['nodes']):
-        props['nodes'][i]['x'] = np.around(props['nodes'][i]['x'],2)
-        props['nodes'][i]['y'] = np.around(props['nodes'][i]['y'],2)
+        props['nodes'][i]['x'] = np.around(props['nodes'][i]['x'],nprec)
+        props['nodes'][i]['y'] = np.around(props['nodes'][i]['y'],nprec)
+        props['nodes'][i]['x_canvas'] = np.around(props['nodes'][i]['x_canvas'],nprec)
+        props['nodes'][i]['y_canvas'] = np.around(props['nodes'][i]['y_canvas'],nprec)
+
+def _assert_positions_within_one_percent(props1,props2):
+
+    for i, n in enumerate(props1['nodes']):
+        assert(np.isclose(props1['nodes'][i]['x'], props2['nodes'][i]['x'],rtol=1e-2))
+        assert(np.isclose(props1['nodes'][i]['y'], props2['nodes'][i]['y'],rtol=1e-2))
+        assert(np.isclose(props1['nodes'][i]['x_canvas'], props2['nodes'][i]['x_canvas'],rtol=1e-2))
+        assert(np.isclose(props1['nodes'][i]['y_canvas'], props2['nodes'][i]['y_canvas'],rtol=1e-2))
+
 
 class Test(unittest.TestCase):
 
@@ -39,15 +50,17 @@ class Test(unittest.TestCase):
         G = _get_test_network()
         props, config = visualize(G,config=_get_test_config(),is_test=True)        
 
-        bind_positions_to_network(G, props)
+        bind_properties_to_network(G, props)
         newprops, newconfig = visualize(G, config=config,is_test=True)
 
-        # round all positional values to their 4th decimal place
-        _round_positions(props)
-        _round_positions(newprops)
+        # test if node positions are close within 1% and subsequently 
+        # round all positional values to the 2nd positon (e.g. 451 => 450)
+        _assert_positions_within_one_percent(props,newprops)
+        _drastically_round_positions(props)
+        _drastically_round_positions(newprops)
 
         # the second visualization should have frozen the nodes
-        assert(newconfig['freeze_nodes'] == True)
+        assert(newconfig['freeze_nodes'])
 
         # change it so the dictionary-equal-assertion doesnt fail
         newconfig['freeze_nodes'] = False
@@ -76,72 +89,52 @@ class Test(unittest.TestCase):
 
     def test_matplotlib(self):
         """Test how the produced figure looks"""
-        stylized_network = {   'linkAlpha': 0.5,
-                               'linkColor': '#7c7c7c',
-                               'links': [   {'source': 0, 'target': 1, 'width': 8.067180430164552},
-                                            {'source': 0, 'target': 2, 'width': 8.067180430164552},
-                                            {'source': 0, 'target': 4, 'width': 8.067180430164552},
-                                            {'source': 0, 'target': 6, 'width': 8.067180430164552},
-                                            {'source': 1, 'target': 3, 'width': 8.067180430164552},
-                                            {'source': 1, 'target': 7, 'width': 8.067180430164552},
[…]
