+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.1.0] -2019-05-23
+### Changed
+- `netwulf.tools.bind_positions_to_network` is now called `netwulf.bind_properties_to_network` and it now does exactly that -- write properties instead of just positions
+- several visualization config options were renamed
+- `netwulf.tools.draw_netwulf` now draws the canvas positions instead of the actual node positions (differences arise by zooming).
+
 ## [v0.0.18] - 2019-05-15
 ### Added
 - added automated test functionality in ``/tests/`` and ``Makefile``
@@ -18,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - switched from usage of `os.path` to `pathlib` at the appropriate places
 
-[Unreleased]: https://github.com/benmaier/netwulf/compare/v0.0.18...HEAD
+[Unreleased]: https://github.com/benmaier/netwulf/compare/v0.1.0...HEAD
+[v0.0.18]: https://github.com/benmaier/netwulf/compare/v0.0.18...v0.1.0
 [v0.0.18]: https://github.com/benmaier/netwulf/compare/v0.0.17...v0.0.18
 [v0.0.17]: https://github.com/benmaier/netwulf/releases/tag/v0.0.17
