### 2019-05-23 e3826ea
some default setting changes
