+++ netwulf/interactive.py
@@ -137,27 +137,27 @@ class NetwulfHTTPRequestHandler(http.server.SimpleHTTPRequestHandler):
 
 default_config = {
     # Input/output
-    'zoom': 1.5,
+    'zoom': 1,
     # Physics
-    'node_charge': -30,
+    'node_charge': -45,
     'node_gravity': 0.1,
-    'link_distance': 10,
+    'link_distance': 15,
     'link_distance_variation': 0,
-    'node_collision': False,
+    'node_collision': True,
     'wiggle_nodes': False,
     'freeze_nodes': False,
     # Nodes
-    'node_fill_color': '#16a085',
-    'node_stroke_color': '#000000',
+    'node_fill_color': '#79aaa0',
+    'node_stroke_color': '#555555',
     'node_label_color': '#000000',
     'display_node_labels': False,
     'scale_node_size_by_strength': False,
-    'node_size': 10,
-    'node_stroke_width': 0.5,
+    'node_size': 5,
+    'node_stroke_width': 1,
     'node_size_variation': 0.5,
     # Links
     'link_color': '#7c7c7c',
-    'link_width': 5,
+    'link_width': 2,
     'link_alpha': 0.5,
     'link_width_variation': 0.5,
     # Thresholding
@@ -335,7 +335,8 @@ def visualize(network,
 
 if __name__ == "__main__":
     # download_d3()
-    G = nx.fast_gnp_random_graph(5,0.3)
-    posted_data = visualize(G,config={'node_size':5,'collision':True,'link_color':'#3b9'},verbose=True)
-    if posted_data is not None:
-        print("received posted data:", posted_data)
+    G = nx.fast_gnp_random_graph(100,2/100.)
+    #G = nx.barabasi_albert_graph(100,1)
+    posted_data = visualize(G,config={'collision':True},verbose=True)
+    #if posted_data is not None:
+    #    print("received posted data:", posted_data)
