### 2019-05-24 c442b3f
paper and outlook in readme
