+++ README.md
@@ -9,6 +9,11 @@ The package is build around the philosophy that network manipulation and preproc
 
 ![example](https://github.com/benmaier/netwulf/raw/master/img/simple_example.gif)
 
+## Paper
+
+[![status](http://joss.theoj.org/papers/3a22c963a45dbddc8501a4b5ef4b2bf6/status.svg)](http://joss.theoj.org/papers/3a22c963a45dbddc8501a4b5ef4b2bf6)
+
+If you use `netwulf` for your scientific work, consider citing us! We're currently [under review in JOSS](https://github.com/openjournals/joss-reviews/issues/1425).
 
 ## Install
 
@@ -61,7 +66,7 @@ This project is licensed under the [MIT License](https://github.com/benmaier/net
 
 ## Contributing
 
-If you want to contribute to this project, please make sure to read the [code of conduct](https://github.com/benmaier/netwulf/blob/master/CODE_OF_CONDUCT.md) and the [contributing guidelines](https://github.com/benmaier/netwulf/blob/master/CONTRIBUTING.md).
+If you want to contribute to this project, please make sure to read the [code of conduct](https://github.com/benmaier/netwulf/blob/master/CODE_OF_CONDUCT.md) and the [contributing guidelines](https://github.com/benmaier/netwulf/blob/master/CONTRIBUTING.md). We're always collecting ideas of what we want to implement next in the [outlook notes](ttps://github.com/benmaier/netwulf/blob/master/OUTLOOK.md).
 
 [![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-v1.4%20adopted-ff69b4.svg)](code-of-conduct.md)
 
