### 2019-05-24 0257d6d
adjusted dpi for matplotlib reproduction
