+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v0.1.1] - 2019-05-24
 ### Changed
 - some default settings
+- set constant `dpi = 72` for unit conversions in matplotlib redrawing because apparently matplotlib only uses this value for conversions, see https://stackoverflow.com/a/35501485/4177832.
 
-## [v0.1.0] -2019-05-23
+## [v0.1.0] - 2019-05-23
 ### Changed
 - `netwulf.tools.bind_positions_to_network` is now called `netwulf.bind_properties_to_network` and it now does exactly that -- write properties instead of just positions
 - several visualization config options were renamed
@@ -26,7 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - switched from usage of `os.path` to `pathlib` at the appropriate places
 
-[Unreleased]: https://github.com/benmaier/netwulf/compare/v0.1.0...HEAD
-[v0.0.18]: https://github.com/benmaier/netwulf/compare/v0.0.18...v0.1.0
+[Unreleased]: https://github.com/benmaier/netwulf/compare/v0.1.1...HEAD
+[v0.1.1]: https://github.com/benmaier/netwulf/compare/v0.1.0...v0.1.1
+[v0.1.0]: https://github.com/benmaier/netwulf/compare/v0.0.18...v0.1.0
 [v0.0.18]: https://github.com/benmaier/netwulf/compare/v0.0.17...v0.0.18
 [v0.0.17]: https://github.com/benmaier/netwulf/releases/tag/v0.0.17
