### 2019-05-25 #1425 @vc1492a
@benmaier Thank you for the detailed replies to my comments. Thank you for writing some automated tests and for providing a structure for writing future tests. The issue of writing tests for an interactive framework is an interesting one, and I think worth exploring in the future as you continue to improve netwulf. I'd very much welcome seeing the addition of curved edges and clustered networks in netwulf as well! Great work.

@leouieda At this stage, I believe we can consider my review of netwulf complete - please let me know if I can further assist in any way.
