### 2019-06-18 72c5b99
fixed some gubs in docs
