### 2019-06-18 4fd3c26
better link zorder behavior
