### 2019-06-21 53c273b
(a) merged js with master; (b) convert numpy types to python types before dumping to json
