### 2019-06-21 5b85b78
stop the server as soon as the browser window is closed
