### 2019-06-26 #1425 @leouieda
@benmaier there shouldn't be a problem. Before acceptance I would ask you to make a release with the changes that came about during review anyway. We'll update the actual version number later on as well. Sorry for the delay with this. I'm trying to find another reviewer that could give me a second opinion.
