### 2019-06-26 #1425 @camillescott
I'm so sorry folks, I let this completely slip my radar. I can get you in a
review by later tonight, unless you'd like to find another person.
Apologies again.

On Tue, Jun 25, 2019 at 7:00 PM Leonardo Uieda <notifications@github.com>
wrote:

> @benmaier <https://github.com/benmaier> there shouldn't be a problem.
> Before acceptance I would ask you to make a release with the changes that
> came about during review anyway. We'll update the actual version number
> later on as well. Sorry for the delay with this. I'm trying to find another
> reviewer that could give me a second opinion.
>
