### 2019-08-14 #1425 @leouieda
@benmaier I've been having trouble getting a second reviewer for this so I'll have a go myself. Should be straight forward since I've already had a look at the package and paper. Sorry for the delay.
