### 2019-08-31 #1425 @leouieda
@benmaier I have finished my review of the software and I'm very happy with all the changes made and the status of the package and paper. Together with the thorough review of @vc1492a this submission if ready for acceptance :tada:

I made a small PR to add a missing DOI (benmaier/netwulf#34). Once that is merged, there are a few actions that you'll need to take:

- Make a release with the latest changes from the review and post the version number here.
- Archive the release on [Zenodo](https://zenodo.org/) and post the DOI here (make sure the title and author list matches the JOSS paper, if possible).

Let me know if you need help with any of these steps.
