### 2019-09-09 a3ba8e3
catch error if multigraph
