### 2019-09-09 #1425 @benmaier
Hi @leouieda, thank you for your editing, review, and pull request! I think I've completed all the necessary steps for publishing:

* I double-checked author-names, affiliations, and ORCIDs.
* I made a new release containing all changes: [v0.1.4](https://github.com/benmaier/netwulf/releases/tag/v0.1.4)
* I archived this [release on Zenodo](https://zenodo.org/record/3402870) (DOI: [10.5281/zenodo.3402870](https://doi.org/10.5281/zenodo.3402870))

Note that I also reformulated a sentence which appeared to be incorrect: In the original paper we claimed that graph-tool does not have interactive features which is evidently false. I therefore rephrased this statement.

Thank you very much, again, and let me know if there's still anything missing!
