### 2019-09-22 #1425 @leouieda
:wave: Hi @benmaier, please change the Graphviz citation to https://link.springer.com/chapter/10.1007/3-540-45848-4_57 (same authors so I'm guessing this is a paper made from the proceedings). I was just going through the paper once more and noticed the citation for the StackOverflow survey :) I couldn't find a proper way of citing it and the current citation isn't rendering correctly with our template. Could you please change it to a hyperlink instead of a citation?
