+++ paper/paper.bib
@@ -59,15 +59,26 @@
     year = {2008}
 }
 
-@inproceedings{graphviz,
-    title = {Graphviz { \textemdash } open source graph drawing tools},
-    booktitle = {Lecture { Notes } in { Computer } { Science }},
-    publisher = {Springer-Verlag},
-    author = {Ellson, John and Gansner, Emden and Koutsofios, Lefteris and North, Stephen and Woodhull, Gordon and Description, Short and Technologies, Lucent},
-    year = {2001},
-    pages = {483--484},
-    file = {Citeseer - Full Text PDF:/Users/bfmaier/Zotero/storage/LS2KQZ99/Ellson et al. - 2001 - Graphviz { \textemdash } open source graph drawing tools.pdf:application/pdf;Citeseer - Snapshot:/Users/bfmaier/Zotero/storage/IXILUSLA/summary.html:text/html}
+@InProceedings{graphviz,
+    author="Ellson, John
+    and Gansner, Emden
+    and Koutsofios, Lefteris
+    and North, Stephen C.
+    and Woodhull, Gordon",
+    editor="Mutzel, Petra
+    and J{\"u}nger, Michael
+    and Leipert, Sebastian",
+    title="Graphviz--- Open Source Graph Drawing Tools",
+    booktitle="Graph Drawing",
+    year="2002",
+    publisher="Springer Berlin Heidelberg",
+    address="Berlin, Heidelberg",
+    pages="483--484",
+    abstract="Graphviz is a heterogeneous collection of graph drawing tools containing batch layout programs (dot, neato, fdp, twopi); a platform for incremental layout (Dynagraph); customizable graph editors (dotty, Grappa); a server for including graphs in Web pages (WebDot); support for graphs as COM objects (Montage); utility programs useful in graph visualization; and libraries for attributed graphs. The software is available under an Open Source license. The article[1] provides a detailed description of the package.",
+    isbn="978-3-540-45848-7"
 }
+
+
 
 @online{developersurvey,
     author = {Stack Overflow},
