### 2019-09-24 #1425 @benmaier
@ulfaslak the Stack overflow reference still looks weird
