+++ paper/paper.bib
@@ -81,8 +81,8 @@
 
 
 
-@misc{developersurvey,
-    address = {https://insights.stackoverflow.com/survey/2019?utm_source=so-owned&utm_medium=blog&utm_campaign=dev-survey-2019&utm_content=launch-blog},
+@manual{developersurvey,
+    url = {https://insights.stackoverflow.com/survey/2019?utm_source=so-owned&utm_medium=blog&utm_campaign=dev-survey-2019&utm_content=launch-blog},
     title = {Stack Overflow Developer Survey},
     year = {2019},
     urldate = {2019-04-23}
