### 2019-09-24 e0998b5
some reference fixes
