+++ paper/paper.bib
@@ -77,12 +77,13 @@
     pages="483--484",
     abstract="Graphviz is a heterogeneous collection of graph drawing tools containing batch layout programs (dot, neato, fdp, twopi); a platform for incremental layout (Dynagraph); customizable graph editors (dotty, Grappa); a server for including graphs in Web pages (WebDot); support for graphs as COM objects (Montage); utility programs useful in graph visualization; and libraries for attributed graphs. The software is available under an Open Source license. The article[1] provides a detailed description of the package.",
     isbn="978-3-540-45848-7"
+    doi={10.1007/3-540-45848-4_57}
 }
 
 
 
 @manual{developersurvey,
-    url = {https://insights.stackoverflow.com/survey/2019?utm_source=so-owned&utm_medium=blog&utm_campaign=dev-survey-2019&utm_content=launch-blog},
+    url = {https://insights.stackoverflow.com/survey/2019},
     title = {Stack Overflow Developer Survey},
     year = {2019},
     urldate = {2019-04-23}
