### 2019-09-24 571e981
some reference fixes
