### 2019-09-30 #1425 @leouieda
@openjournals/joss-eics this submission is ready of acceptance :confetti_ball:
