### 2019-10-01 #1425 @kyleniemeyer
Hi @benmaier, I fixed the NetworkX reference, please merge this PR: https://github.com/benmaier/netwulf/pull/35
