### 2019-10-01 #1425 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
